change deposit-list-item to dummy component, resolve #14
