+++ src/dummies/deposit-list-item/deposit-list-item.scss
